Structure and document protocol between server and client
In order to be able to create clients in multiple languages,
we need a universal (de-)serializer.

Although binary marshalling reduces network bandwidth, we decided
not to use it, in favor of readability and debugability, i.e.
being able to use tc to send request to server is pretty useful
and powerful

We use simple protocol to send reponses and notifications as well.

Security consideration of text-based protocol are out of the scope of this
changes, as now we are not sure what level of security we want to have.

Signed-off-by: Pavel Abramov <uncle.decart@gmail.com>