Add trace logging library and basic server-side code logging
logging is crucial to debug any issues that one might encounter in production
or development stage. And it is crucial to get logging good enough.

This patch introcudes tracing library to nkv system for following reasons:

1) It has reach functionality which allows to log asynchronous code correctly
2) It supports multiple outputs for logs, such as files, stdout and we can expand
it to support custom sinks, like Unix Domain Socket for memlogd
3) It supports sending logs in OpenTelemetry (OTel) format, which is at the moment
standard de facto for metrics processing, they also support logging processing,
it is better to use a standard than to re-invent a wheel

Signed-off-by: Pavel Abramov <uncle.decart@gmail.com>